# rpc: implement DecompressedSize on snappyCompressor