… pkg/rpc/snappy_test.go