This commit updates `snappyCompressor` to implement the anonymous interface declared in `grpc/rpc_util.go`'s decompress function. When implemented by an `encoding.Compressor`, the function can pre-allocate a buffer with the exact decompressed size instead of needing to dynamically grow the buffer while decompressing (in io.ReadAll), which can incur multiple allocations and memory copies.

This optimization has been available since https://github.com/grpc/grpc-go/commit/dcd1c9748d337a77f783b8be8666aa3d6fd47373.

See https://cockroachlabs.slack.com/archives/C4X2J0RH6/p1671224064592639 for discussion.

```
➜ benchdiff --run='BenchmarkReplicaProposal/bytes=.*,withFollower=true' --count=25 --old=HEAD~2 ./pkg/kv/kvserver

name                                                old time/op    new time/op    delta
ReplicaProposal/bytes=1.0_MiB,withFollower=true-10    3.23ms ± 6%    3.11ms ± 5%  -3.60%  (p=0.000 n=25+23)
ReplicaProposal/bytes=256_KiB,withFollower=true-10     835µs ± 7%     808µs ± 8%  -3.16%  (p=0.003 n=25+24)
ReplicaProposal/bytes=512_KiB,withFollower=true-10    1.67ms ± 6%    1.63ms ± 4%  -2.50%  (p=0.001 n=24+24)
ReplicaProposal/bytes=2.0_MiB,withFollower=true-10    6.17ms ± 9%    6.04ms ± 6%  -2.15%  (p=0.026 n=25+24)
ReplicaProposal/bytes=256_B,withFollower=true-10      44.9µs ± 2%    44.9µs ± 4%    ~     (p=0.945 n=24+25)
ReplicaProposal/bytes=512_B,withFollower=true-10      46.9µs ± 5%    46.9µs ± 3%    ~     (p=0.562 n=25+24)
ReplicaProposal/bytes=1.0_KiB,withFollower=true-10    49.3µs ± 4%    49.2µs ± 3%    ~     (p=0.835 n=25+24)

name                                                old speed      new speed      delta
ReplicaProposal/bytes=1.0_MiB,withFollower=true-10   325MB/s ± 6%   336MB/s ± 7%  +3.39%  (p=0.000 n=25+24)
ReplicaProposal/bytes=256_KiB,withFollower=true-10   314MB/s ± 7%   325MB/s ± 8%  +3.28%  (p=0.003 n=25+24)
ReplicaProposal/bytes=512_KiB,withFollower=true-10   315MB/s ± 6%   323MB/s ± 4%  +2.55%  (p=0.001 n=24+24)
ReplicaProposal/bytes=2.0_MiB,withFollower=true-10   341MB/s ± 9%   348MB/s ± 6%  +2.13%  (p=0.026 n=25+24)
ReplicaProposal/bytes=256_B,withFollower=true-10    5.70MB/s ± 2%  5.70MB/s ± 4%    ~     (p=0.941 n=24+25)
ReplicaProposal/bytes=512_B,withFollower=true-10    10.9MB/s ± 5%  10.9MB/s ± 3%    ~     (p=0.562 n=25+24)
ReplicaProposal/bytes=1.0_KiB,withFollower=true-10  20.8MB/s ± 4%  20.8MB/s ± 3%    ~     (p=0.847 n=25+24)

name                                                old alloc/op   new alloc/op   delta
ReplicaProposal/bytes=2.0_MiB,withFollower=true-10     102MB ± 4%      92MB ± 4%  -9.42%  (p=0.000 n=25+25)
ReplicaProposal/bytes=1.0_MiB,withFollower=true-10    55.8MB ± 2%    50.7MB ± 4%  -9.27%  (p=0.000 n=21+25)
ReplicaProposal/bytes=256_KiB,withFollower=true-10    14.8MB ± 4%    13.6MB ± 4%  -8.24%  (p=0.000 n=25+24)
ReplicaProposal/bytes=512_KiB,withFollower=true-10    27.4MB ± 3%    25.3MB ± 4%  -7.52%  (p=0.000 n=23+25)
ReplicaProposal/bytes=1.0_KiB,withFollower=true-10    53.6kB ± 2%    52.8kB ± 2%  -1.60%  (p=0.000 n=25+25)
ReplicaProposal/bytes=512_B,withFollower=true-10      35.3kB ± 1%    35.5kB ± 1%  +0.65%  (p=0.000 n=25+24)
ReplicaProposal/bytes=256_B,withFollower=true-10      26.3kB ± 2%    27.0kB ± 1%  +2.79%  (p=0.000 n=25+25)

name                                                old allocs/op  new allocs/op  delta
ReplicaProposal/bytes=512_KiB,withFollower=true-10     1.83k ±10%     1.76k ± 5%  -4.05%  (p=0.002 n=25+23)
ReplicaProposal/bytes=256_KiB,withFollower=true-10     1.02k ±10%     0.98k ± 8%  -4.00%  (p=0.001 n=25+24)
ReplicaProposal/bytes=1.0_MiB,withFollower=true-10     3.53k ± 7%     3.40k ± 5%  -3.66%  (p=0.002 n=25+22)
ReplicaProposal/bytes=1.0_KiB,withFollower=true-10       145 ± 0%       143 ± 2%  -1.24%  (p=0.000 n=18+24)
ReplicaProposal/bytes=512_B,withFollower=true-10         142 ± 1%       141 ± 1%  -0.82%  (p=0.000 n=24+24)
ReplicaProposal/bytes=256_B,withFollower=true-10         139 ± 1%       139 ± 1%    ~     (p=0.694 n=25+24)
ReplicaProposal/bytes=2.0_MiB,withFollower=true-10     6.80k ±12%     6.65k ±14%    ~     (p=0.247 n=25+25)
```

Release note (performance improvement): The RPC compression encoding was extended with a length prefixing format, allowing more efficient decompression on receivers.

Epic: None